# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * WCAG conformance level to test against
 */
export type ScanRequestWcagLevel =
  (typeof ScanRequestWcagLevel)[keyof typeof ScanRequestWcagLevel];

export const ScanRequestWcagLevel = {
  A: "A",
  AA: "AA",
  AAA: "AAA",
} as const;

export interface ScanRequest {
  /** URL to scan for accessibility issues */
  url: string;
  /** WCAG conformance level to test against */
  wcagLevel?: ScanRequestWcagLevel;
  /** Include WCAG 2.2 rules */
  includeWcag22?: boolean;
  /** Include AAA-level rules */
  includeAaa?: boolean;
}

export interface A11yNode {
  html: string;
  target: string[];
  failureSummary?: string | null;
}

export interface A11yViolation {
  id: string;
  impact?: string | null;
  description: string;
  help: string;
  helpUrl: string;
  tags: string[];
  nodes: A11yNode[];
}

export interface A11yRule {
  id: string;
  description: string;
  help: string;
  helpUrl: string;
  tags: string[];
}

export interface ScanSummary {
  violationCount: number;
  passCount: number;
  incompleteCount: number;
  criticalCount: number;
  seriousCount: number;
  moderateCount: number;
  minorCount: number;
}

export interface ScanResult {
  url: string;
  /** Scan duration in milliseconds */
  scanDuration: number;
  timestamp: string;
  violations: A11yViolation[];
  passes: A11yRule[];
  incomplete: A11yRule[];
  inapplicable: A11yRule[];
  summary: ScanSummary;
}

export interface ErrorResponse {
  error: string;
  message: string;
}