# lib/api-zod/src/generated/types/a11yNode.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface A11yNode {
  html: string;
  target: string[];
  failureSummary?: string | null;
}