# lib/api-zod/src/generated/types/a11yRule.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface A11yRule {
  id: string;
  description: string;
  help: string;
  helpUrl: string;
  tags: string[];
}