# lib/api-zod/src/generated/types/a11yViolation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { A11yNode } from "./a11yNode";

export interface A11yViolation {
  id: string;
  impact?: string | null;
  description: string;
  help: string;
  helpUrl: string;
  tags: string[];
  nodes: A11yNode[];
}