# lib/api-zod/src/generated/types/scanRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScanRequestWcagLevel } from "./scanRequestWcagLevel";

export interface ScanRequest {
  /** URL to scan for accessibility issues */
  url: string;
  /** WCAG conformance level to test against */
  wcagLevel?: ScanRequestWcagLevel;
  /** Include WCAG 2.2 rules */
  includeWcag22?: boolean;
  /** Include AAA-level rules */
  includeAaa?: boolean;
}