# lib/api-zod/src/generated/types/scanResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { A11yRule } from "./a11yRule";
import type { A11yViolation } from "./a11yViolation";
import type { ScanSummary } from "./scanSummary";

export interface ScanResult {
  url: string;
  /** Scan duration in milliseconds */
  scanDuration: number;
  timestamp: string;
  violations: A11yViolation[];
  passes: A11yRule[];
  incomplete: A11yRule[];
  inapplicable: A11yRule[];
  summary: ScanSummary;
}